# [ZOrder] Fast approach of interleaving bits