--- a/core/src/main/scala/org/apache/spark/sql/delta/sources/DeltaSQLConf.scala
+++ b/core/src/main/scala/org/apache/spark/sql/delta/sources/DeltaSQLConf.scala
@@ -727,6 +727,7 @@ trait DeltaSQLConfBase {
       .booleanConf
       .createWithDefault(true)
 
+<<<<<<< HEAD
   val DELTA_CDF_ALLOW_OUT_OF_RANGE_TIMESTAMP = {
     buildConf("changeDataFeed.timestampOutOfRange.enabled")
       .doc(
@@ -746,6 +747,16 @@ trait DeltaSQLConfBase {
       .internal()
       .booleanConf
       .createWithDefault(true)
+
+  val FAST_INTERLEAVE_BITS_ENABLED =
+    buildConf("fast.interleaveBits.enabled")
+      .internal()
+      .doc(
+        """
+          |When true using the O(4 * 8) algorithm for interleave bit, otherwise the O(32 * n)
+          |algorithm, n is the number of input columns""".stripMargin)
+      .booleanConf
+      .createWithDefault(false)
 }
 
 object DeltaSQLConf extends DeltaSQLConfBase