<!--
Thanks for sending a pull request!  Here are some tips for you:
  1. If this is your first time, please read our contributor guidelines: https://github.com/delta-io/delta/blob/master/CONTRIBUTING.md
  2. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP] Your PR title ...'.
  3. Be sure to keep the PR description updated to reflect all changes.
  4. Please write your PR title to summarize what this PR proposes.
  5. If possible, provide a concise example to reproduce the issue for a faster review.
  6. If applicable, include the corresponding issue number in the PR title and link it in the body.
-->

## Description
to resolve https://github.com/delta-io/delta/pull/1149#discussion_r881782878

Improve interleave bits perf, the basic idea is from http://graphics.stanford.edu/~seander/bithacks.html#InterleaveTableObvious


## How was this patch tested?

<!--
If tests were added, say they were added here. Please make sure to test the changes thoroughly including negative and positive cases if possible.
If the changes were tested in any way other than unit tests, please clarify how you tested step by step (ideally copy and paste-able, so that other reviewers can test and check, and descendants can verify in the future).
If the changes were not tested, please explain why.
-->

Existing UTs.

## Does this PR introduce _any_ user-facing changes?

<!--
If yes, please clarify the previous behavior and the change this PR proposes - provide the console output, description and/or an example to show the behavior difference if possible.
If possible, please also clarify if this is a user-facing change compared to the released Delta Lake versions or within the unreleased branches such as master.
If no, write 'No'.
-->

No.

# Benchmark

Before this PR:

```
[info] OpenJDK 64-Bit Server VM 1.8.0_292-8u292-b10-0ubuntu1~18.04-b10 on Linux 5.4.0-110-generic
[info] Intel(R) Xeon(R) CPU E5-2660 v4@ 2.00GHz
[info] 1000000 rows interleave bits benchmark:   Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
[info] ------------------------------------------------------------------------------------------------------------------------
[info] 1 int columns benchmark                             590            612          27          1.7         590.0       1.0X
[info] 2 int columns benchmark                             858            862           8          1.2         857.8       0.7X
[info] 3 int columns benchmark                            1102           1121          29          0.9        1102.5       0.5X
[info] 4 int columns benchmark                            1266           1291          23          0.8        1265.7       0.5X
```

After this PR:
```
[info] OpenJDK 64-Bit Server VM 1.8.0_292-8u292-b10-0ubuntu1~18.04-b10 on Linux 5.4.0-110-generic
[info] Intel(R) Xeon(R) CPU E5-2660 v4@ 2.00GHz
[info] 1000000 rows interleave bits benchmark:   Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
[info] ------------------------------------------------------------------------------------------------------------------------
[info] 1 int columns benchmark                             454            473          21          2.2         454.1       1.0X
[info] 2 int columns benchmark                             633            646          12          1.6         633.0       0.7X
[info] 3 int columns benchmark                             821            824           4          1.2         820.9       0.6X
[info] 4 int columns benchmark                             966            987          30          1.0         965.6       0.5X
```