@tdas 2022-05-31
@cfmcgrady This is absolutely great addition. I took a very quick cursory look, not in the details of the new logic, but overall PR. My major concern is that this is a lot of new code with very custom logic for handling up to some N columns in a more optimized way, but there isnt a lot of unit tests to really ensure the test coverage of the correctness in these cases. The baseline default case has a simple logic and therefore simple to test. All of these new code paths need sufficient testing. However, it also hard to explicitly conclusively test every combination that have been handled specifically. So here are a bunch suggestions that will improve my confidence in the PR. 

1. Put this new code path behind a flag in DeltaSQLConf so that we can easily turn if off if there are concerns. 

2. For test coverage, lets try to do randomized testing. Generate a sequence of random int, run it through the default and fast path, and make sure that the result is the same. 
  - Make sure that the randomization is based on a randomly picked seed that is printed in the logs, so that if there is an error, we can use the logged seed to reproduce the exact input and output. 
  - Make the number of iterations configurable in the code so that we can run it manually like 1B times (will take a while), but when we check it in, we do it like less iterations that will take say < 10 seconds.
 
what do you think?