@cfmcgrady 2022-06-06
> Put this new code path behind a flag in DeltaSQLConf so that we can easily turn if off if there are concerns.

Added a new configuration `fast.interleaveBits.enabled`, which is enabled by default.

> For test coverage, lets try to do randomized testing. Generate a sequence of random int, run it through the default and fast path, and make sure that the result is the same.

Added new randomized testing, which runs 100,000,000 random interleaved bit test cases.

```
[info] InterleaveBitsSuite:
[info] - 0 inputs (85 milliseconds)
[info] - 1 input (304 milliseconds)
[info] - 2 inputs (26 milliseconds)
[info] - 3 inputs (30 milliseconds)
[info] - nulls (221 milliseconds)
[info] - consistency (253 milliseconds)
[info] - supported types (11 milliseconds)
[info] - randomization interleave bits (2 minutes, 29 seconds)
```