@felipepessoto 2022-12-08
> > Put this new code path behind a flag in DeltaSQLConf so that we can easily turn if off if there are concerns.
> 
> Added a new configuration `spark.databricks.delta.fast.interleaveBits.enabled`, which is enabled by default.
> 

@cfmcgrady, The config is actually disabled by default, is that expected?
cc @tdas